requirements.txt
aerich==0.9.1
aiofiles==24.1.0
aiomysql==0.2.0
aiosmtplib==5.0.0
aiosqlite==0.22.0
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
asyncclick==8.3.0.7
atlastk==0.13.5
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dictdiffer==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
Faker==37.12.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
fastapi-mail==1.6.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iso8601==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pypika-tortoise==0.6.3
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
slugify==0.0.1
starlette==0.48.0
tortoise-orm==0.25.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
